Fix time printing + switch timestamps to unsigned int

+++ philo_one/includes/philo_one.h
@@ -6,7 +6,7 @@
 /*   By: sadolph <sadolph@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/21 15:45:38 by sadolph           #+#    #+#             */
-/*   Updated: 2021/01/14 18:52:08 by sadolph          ###   ########.fr       */
+/*   Updated: 2021/01/15 14:15:55 by sadolph          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,7 @@ typedef struct		s_table
 	int				die;
 	int				eat;
 	int				sleep;
-	long			start;
+	unsigned int	start;
 	pthread_mutex_t	*print;
 	pthread_mutex_t	*time;
 	pthread_mutex_t	*satisfied;
@@ -45,7 +45,7 @@ typedef struct		s_table
 typedef struct		s_philo
 {
 	int				id;
-	long 			last_eat;
+	unsigned int	last_eat;
 	t_table			*table;
 	pthread_mutex_t	*fork_l;
 	pthread_mutex_t	*fork_r;
